Inline some clock functions
